update: Homeコンポーネントをサーバーコンポーネントにし、PopulationChartとして切り出し
